+++ engines/text/duckduckgo.php
